separate end points to be disabled
